+++ src/PICHexDisassembler/Instructions/Bsf.cs
@@ -2,7 +2,7 @@
 {
     public class Bsf : BitInstruction
     {
-        public Bsf(int data) : base(data)
+        public Bsf(ushort data) : base(data)
         {
         }
     }
